requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
bidict==0.23.1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
decorator==4.4.2
descope==1.6.7
distro==1.9.0
dnspython==2.6.1
docutils==0.21.2
email_validator==2.2.0
fastapi==0.109.2
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-core==2.4.1
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ifaddr==0.2.0
imageio==2.35.1
imageio-ffmpeg==0.4.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
liccheck==0.9.2
markdown2==2.5.0
MarkupSafe==2.1.5
moviepy==1.0.3
multidict==6.0.5
mutagen==1.46.0
nicegui==1.4.36
numpy==2.1.0
openai==1.42.0
orjson==3.10.7
pillow==10.4.0
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.28.0
pscript==0.7.7
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
semantic-version==2.10.0
setuptools==73.0.1
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.36.3
stripe==10.10.0
toml==0.10.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
vbuild==0.8.2
watchfiles==0.23.0
websockets==13.0
wsproto==1.2.0
yarl==1.9.4
pydub==0.25.1